Create qcpConsoleLog
Provides quoteLine, quote and conn model variables available in Salesforce CPQ's calculation engine used in a Quote Calculator Plug-in.  Put this code in the Code field for a Custom Script record and enter the name of the Script in the Installed Package settings for Salesforce CPQ under Plug-ins.  Navigate to a Quote record, click Edit Lines, right click and inspect in the browser's console (NOT dev console) and click calculate.  All the variables should print.