qcpConsoleLog
Performs a JavaScript print action in the web browser for Quote, Quote Line and Connection objects.  This code should be copied and pasted in the Custom Script record in your Salesforce CPQ org.  While on the Quote Line Editor (accessed via Quote record -> Edit Lines), right click to inspect page, click Calculate and view the results.  

This QCP function is particularly useful to display the model's attributes that are loaded in for use when developing advanced CPQ pricing functionality.