Update QCP_CalculateYearOverYearFinancials
Sums yearly totals for MDQ products, calculates financials based off previous year, checks for duplicate "Grouping" quote lines.  A grouping is a product record that is not sold to customer, but acts as a Product Grouping where like-products are selected.  This is primarily used to validate the products under the grouping (i.e. exclusions, dependencies, etc.).